Add ShouldRemoveNullStone
This property specifies if the ability should be used to remove a target's Null Stone effect. Abilities like Armadon's Snot or Deadwood's Uproot are perfect candidates for this due to their short cooldowns and low impact (especially late game by the time one would be able to get a Null Stone).